[automerger skipped] Check LAUNCHED_BY_ACLOUD!=true env var then delete the home directory. am: 00ef085df8 am: f6bb6ddafc -s ours
am skip reason: Merged-In If1e679498e88e09af7fbae7585f79d7ac298fe74 with SHA-1 997ed31859 is already in history

Original change: https://googleplex-android-review.googlesource.com/c/device/google/cuttlefish/+/23260841

+++ base/cvd/cuttlefish/host/commands/cvd/server_command/start.cpp
@@ -177,16 +177,16 @@ Result<void> CvdStartCommandHandler::AcloudCompatActions(
     if (!FileExists(acloud_compat_home)) {
       continue;
     }
-    if (!DirectoryExists(acloud_compat_home, /*follow_symlinks=*/false)) {
-      // cvd created a symbolic link
-      result_deleted = RemoveFile(acloud_compat_home);
-    } else {
-      // acloud created a directory
-      // rm -fr isn't supporetd by TreeHugger, so if we fork-and-exec to
-      // literally run "rm -fr", the presubmit testing may fail if ever this
-      // code is tested in the future.
-      if (!Contains(group_creation_info.envs, kLaunchedByAcloud) ||
+    if (!Contains(group_creation_info.envs, kLaunchedByAcloud) ||
         group_creation_info.envs.at(kLaunchedByAcloud) != "true") {
+      if (!DirectoryExists(acloud_compat_home, /*follow_symlinks=*/false)) {
+        // cvd created a symbolic link
+        result_deleted = RemoveFile(acloud_compat_home);
+      } else {
+        // acloud created a directory
+        // rm -fr isn't supporetd by TreeHugger, so if we fork-and-exec to
+        // literally run "rm -fr", the presubmit testing may fail if ever this
+        // code is tested in the future.
         result_deleted = RecursivelyRemoveDirectory(acloud_compat_home);
       }
     }
